Craete crawling_tool : 크롤링 모듈을 위한 함수들을 모아두었습니다
# get_gall_id() : 갤 id 리턴
# get_gall_type() : 메이저갤러리인지 마이너갤러리인지 미니갤러리인지 판단
# get_max_num() : 검색결과 중 가장 큰 글번호를 구하여 리턴한다
# get_last_page() : 검색결과의 마지막 페이지가 몇인지 리턴

requirements.txt 에 webdriver_manger를 추가함
